bing_scrape/requirements.txt
absl-py==1.2.0
alexa-client==1.5.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.1
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi @ file:///C:/ci_310/cffi_1642682485096/work
chardet==5.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
comtypes==1.1.14
conda==4.13.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1649106011304/work
cookiecutter==2.1.1
cryptography @ file:///C:/ci/cryptography_1639414720302/work
cvlib==0.2.7
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==22.9.24
fonttools==4.33.3
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h2==2.6.2
h5py==3.7.0
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.22.0
importlib-metadata==4.12.0
imutils==0.5.4
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
menuinst @ file:///C:/ci/menuinst_1631733438520/work
-e git+https://github.com/projectmesa/mesa@60bc768f5eede2dc4806684ae4b0b3046eb07d27#egg=Mesa
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy @ file:///D:/bld/numpy_1657483989399/work
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas @ file:///D:/bld/pandas_1656001279371/work
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
progressbar==2.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.5
psutil==5.9.1
pure-eval==0.2.2
py4j @ file:///home/conda/feedstock_root/build_artifacts/py4j_1647581994053/work
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pyspark @ file:///home/conda/feedstock_root/build_artifacts/pyspark_1655363664545/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-decouple==3.6
python-slugify==6.1.2
pyttsx3==2.90
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1647961439546/work
pywin32==302
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.2.0
qrcode==7.3.1
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
resettabletimer==0.7.0
rsa==4.9
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==5.3.0
typing_extensions==4.3.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
